prints elements in liked list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include <stddef.h>
 #include "lists.h"
 
 /**
@@ -12,7 +11,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned int i;
+	size_t i;
 
 	for (i = 0; h; i++)
 	{
